+++ app/certifications-on-resume/Content.tsx
@@ -1,5 +1,5 @@
 'use client';
-/* eslint-disable react/no-unescaped-entities */
+ 
 
 import BlogPostLayout from '@/components/BlogPostLayout';
 import { useLoginGateway } from '@/components/LoginGateway';